requirements.txt
pandas
twitterscrapper
googlesearch-python
csv
